Remove default socket timeout value

+++ components/org.wso2.transport.http.netty/src/main/java/org/wso2/transport/http/netty/listener/ServerConnectorBootstrap.java
@@ -128,7 +128,7 @@ public class ServerConnectorBootstrap {
         }
     }
 
-    public void addIdleTimeout(int socketIdleTimeout) {
+    public void addIdleTimeout(long socketIdleTimeout) {
         httpServerChannelInitializer.setIdleTimeout(socketIdleTimeout);
     }
